requirements.txt
fastapi==0.109.0
uvicorn==0.26.0
Mastodon.py==1.8.1
python-dotenv==1.0.0
